Create torrio.py
This code  reads and parse routine files  to be downloaded. Where an object is created to track  list of peers
file, and Queue created for saving file and writing them on a hard drive and also a connection to keep tracks of communication. before it gets check if data to be downloaded is already there. if downloaded then it gets piece of data downloaded from the data block if not then it it creates a new block for the data to be downloaded and stored in a block for it to be downloaded by others.